check/tree/450/450.删除二叉搜索树中的节点_test.go
package main

import (
	. "leetcode/check/tree"
	"reflect"
	"testing"
)

func Test_deleteNode(t *testing.T) {
	type args struct {
		root *TreeNode
		key  int
	}
	tests := []struct {
		name string
		args args
		want *TreeNode
	}{
		{
			name: "test1",
			args: args{
				key:  50,
				root: BuildStrTree("[50,30,70,null,40,60,80]"),
			},
			want: BuildStrTree("[60,30,70,null,40,null,80]"),
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := deleteNode(tt.args.root, tt.args.key); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("deleteNode() = %v, want %v", got, tt.want)
			}
		})
	}
}
